:recycle: Separate file for json encoder and decoder
